Convert find_local_council integration tests
- local_links_manager_helpers only has one method
  only used in one place, so move it to there.
- First appearance of content_store_has_random_item,
  so bring it into content_store_helper
- bring in location_helpers (first use)
