+++ spec/support/location_helpers.rb
