+++ test/support/content_store_helpers.rb
@@ -1,4 +1,3 @@
-require "govuk_schemas"
 require "gds_api/test_helpers/content_store"
 
 module ContentStoreHelpers
@@ -22,21 +21,6 @@ module ContentStoreHelpers
       "title" => "A Taxon",
       "base_path" => "/a-taxon",
     }
-  end
-
-  def content_store_has_random_item(base_path:, schema: "generic", is_tagged_to_taxon: false, details: {})
-    content_item = GovukSchemas::RandomExample.for_schema(frontend_schema: schema) do |item|
-      taxons = is_tagged_to_taxon ? [basic_taxon] : []
-      item.merge!(
-        "base_path" => base_path,
-        "links" => { "taxons" => taxons },
-      )
-      item["details"].merge!(details)
-      item
-    end
-
-    stub_content_store_has_item(content_item["base_path"], content_item)
-    content_item
   end
 
   def stub_content_store_has_item_tagged_to_taxon(base_path:, payload:)
